Consolidated CSS cleanup and extracted score bar and set bar styling calcuations
