+++ frontend/src/components/AnnotationView/AnnotationSidebar.vue
@@ -28,12 +28,10 @@ export default {
   aside {
     display: flex;
     flex-direction: column;
-    /* justify-content: flex-start; */
     background-color: var(--rosalution-white);
     border-radius: var(--content-border-radius);
     padding: var(--p-8);
     gap: var(--p-8);
-    /* margin-top: var(--p-10); */
     overflow-x: hidden;
   }
 </style>
