+++ frontend/src/components/AnnotationView/SetDataset.vue
@@ -4,12 +4,8 @@
     <span v-if="!availableData" class="set-section dataset-bar dataset-bar-fill-unavailable" data-test="dataset-bar">
     </span>
     <span v-else class="set-section dataset-bar" data-test="dataset-bar">
-      <span v-for="(item, index) in set"
-        :key="`${value}-${index}`"
-        :style="setItemStyle(item)"
-        class="set-text"
-      >
-        {{ classificationText(item) }}
+      <span v-for="item in setItems" :key=item.key :style=item.style class="set-text">
+        {{ item.content }}
       </span>
     </span>
   </div>
@@ -18,7 +14,7 @@
 <script setup>
 import DatasetLabel from '@/components/AnnotationView/DatasetLabel.vue';
 
-import {isDatasetAvailable} from '@/components/AnnotationView/datasetRenderingUtility.js';
+import {isDatasetAvailable, useSetBarCalculations} from '@/components/AnnotationView/datasetRenderingUtility.js';
 
 const props = defineProps({
   label: {
@@ -40,39 +36,6 @@ const props = defineProps({
 });
 
 const availableData = isDatasetAvailable(props.value);
+const setItems = useSetBarCalculations(props.value, props.set);
 
-function calculateWidth(value) {
-  const width = 100;
-
-  if (value == props.value) {
-    return `${width * .75}%`;
-  }
-  return `${ (width / props.set.length) / 2}%`;
-};
-
-function setItemStyle(item) {
-  const style = {
-    'background-color': 'var(--rosalution-grey-300)',
-    'opacity': 1,
-  };
-
-
-  if (item.colour == 'Red') {
-    style['background-color'] = 'var(--rosalution-red-200)';
-  } else if (item.colour == 'Blue') {
-    style['background-color'] = 'var(--rosalution-blue-200)';
-  } else if (item.colour == 'Yellow') {
-    style['background-color'] = 'var(--rosalution-yellow-200)';
-  } else if (item.colour == 'Green') {
-    style['background-color'] = 'var(--rosalution-green-200)';
-  }
-
-  style['width'] = calculateWidth(item.value);
-
-  return (item.value == props.value) ? style : {...style, 'opacity': 0.5};
-};
-
-function classificationText(setItem) {
-  return (setItem.value == props.value) ? setItem.classification : '';
-};
 </script>
